+++ tests/TransportManagerTest.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Coconuts\Mail;
+
+class TransportManagerTest extends TestCase
+{
+    /** @test */
+    public function can_create_postmark_driver()
+    {
+        $manager = new TransportManager($this->app);
+
+        $transport = $manager->createPostmarkDriver();
+
+        $this->assertEquals(PostmarkTransport::class, get_class($transport));
+    }
+}