+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
+## [Unreleased]
+
+## [0.1.0] - 2017-01-29
+
+### Added
+- Adds service provider, manager and mail driver with appropriate tests [`b3c10cd221`](https://github.com/coconutcraig/laravel-postmark/commit/b3c10cd221) 
+
 ## 0.0.1 - 2017-01-29
 
 ### Added
-- Adds package skeleton [`2f6fe84bcc`](https://github.com/coconutcraig/laravel-postmark/commit/2f6fe84bcc) 
+- Adds package skeleton [`2f6fe84bcc`](https://github.com/coconutcraig/laravel-postmark/commit/2f6fe84bcc)
+
+[Unreleased]: https://github.com/coconutcraig/laravel-postmark/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/coconutcraig/laravel-postmark/compare/v0.0.1...v0.1.0